app/requirements.txt
alembic==1.17.2
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
greenlet==3.3.0
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
playwright==1.57.0
pluggy==1.6.0
psycopg2==2.9.11
pyee==13.0.0
Pygments==2.19.2
pytest==9.0.2
pytest-base-url==2.1.0
pytest-playwright==0.7.2
python-dotenv==1.2.1
python-slugify==8.0.4
requests==2.32.5
schedule==1.2.2
SQLAlchemy==2.0.45
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.6.2
